feat(): teach hubSearch channels
