+++ packages/common/src/discussions/api/channels/index.ts
@@ -1,0 +1,1 @@
+export * from "./channels";
