+++ packages/common/src/discussions/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./channels";
+export * from "./types";
