+++ packages/common/src/search/_internal/index.ts
@@ -2,3 +2,4 @@ export * from "./portalSearchItems";
 export * from "./hubSearchItems";
 export * from "./portalSearchGroups";
 export * from "./portalSearchUsers";
+export * from "./discussionsSearchChannels";
