+++ packages/common/test/search/_internal/mocks/searchChannelsResponse.ts
@@ -1,0 +1,27 @@
+const response = {
+  items: [
+    {
+      groups: ["cb0ddfc90f4f45b899c076c88d3fdc84"],
+      id: "69d034b4db8d4dd6bc26ac64169bcbfe",
+      access: "private",
+      allowAnonymous: false,
+      allowedReactions: null,
+      allowReaction: true,
+      allowReply: true,
+      blockWords: null,
+      createdAt: "2023-04-10T04:17:30.498Z",
+      creator: "juliana_pa",
+      defaultPostStatus: "approved",
+      editor: "carla_pa",
+      metadata: null,
+      name: null,
+      orgs: ["Xj56SBi2udA78cC9"],
+      softDelete: true,
+      updatedAt: "2023-04-10T04:17:32.704Z",
+    },
+  ],
+  total: 1,
+  nextStart: -1,
+};
+
+export default response;
